+++ src/bible/routes.nim
@@ -1,11 +1,12 @@
 import pkg/prologue
 
 import bible/routes/[
-  verse,
-  chapter,
-  books,
   documents,
+  books,
+  chapters,
+  verses,
 ]
+import bible/routes/api/incAccess
 
 import bible/routes/default/notFound
 
@@ -17,10 +18,13 @@ type
 const
   routesDefinition*: seq[Route] = @[ ## All application routes
     ("", @[
-      pattern("/{doc}/{book}/{chapter}", r_verse, HttpGet, "vesicles"),
+      pattern("/{doc}/{book}/{chapter}", r_verses, HttpGet, "veses"),
       pattern("/{doc}/{book}", r_chapters, HttpGet, "chapters"),
       pattern("/{doc}", r_books, HttpGet, "books"),
       pattern("/", r_documents, HttpGet, "documents"),
+    ]),
+    ("api", @[
+      pattern("/incAccess", r_incAccess, HttpPost, "incAccess"),
     ]),
   ]