+++ src/bible/routes/documents.nim
@@ -3,7 +3,10 @@ from std/strutils import contains
 
 import pkg/prologue
 
-import bible/db/models/document
+import bible/db/models/[
+  document,
+  access
+]
 import bible/routeUtils
 
 import bible/views
@@ -12,4 +15,7 @@ import bible/views/documents
 proc r_documents*(ctx: Context) {.async.} =
   ## All documents
   ctx.forceHttpMethod HttpGet
-  ctx.render documents(getAllDocs())
+  ctx.render(
+    getAccess("", "", 0).accesses,
+    documents(getAllDocs())
+  )