+++ src/bible/views.nim
@@ -7,14 +7,14 @@ import pkg/karax/[
 ]
 
 from bible/config import withConf, appName
-from bible/utils import url
+from bible/utils import assetUrl
 
 type View* = tuple
   name: string
   vnode: VNode
   code: HttpCode
 
-proc render*(ctx: Context; view: View) =
+proc render*(ctx: Context; accesses: int; view: View) =
   ## Renders the karax element
   var vnode: VNode
   withConf:
@@ -24,11 +24,15 @@ proc render*(ctx: Context; view: View) =
         meta(`http-equiv` = "X-UA-Compatible", content = "IE=edge")
         meta(name = "viewport", content = "width=device-width, initial-scale=1.0")
         title: text fmt"{view.name} - {appName}"
-        link(rel = "stylesheet", href = url "/assets/style/third/mvp.min.css")
-        link(rel = "stylesheet", href = url "/assets/style/bible.css")
+        link(rel = "stylesheet", href = assetUrl "style/third/mvp.min.css")
+        link(rel = "stylesheet", href = assetUrl "style/bible.css")
       body:
         tdiv(class = "content"):
           view.vnode
+          if accesses > 0:
+            footer:
+              tdiv(class = "accesses"):
+                text fmt"This page was accessed {accesses} times in this month"
         footer:
           tdiv(class = "contribute"):
             text "This site is open-source! You can contribute at "
@@ -42,6 +46,6 @@ proc render*(ctx: Context; view: View) =
               text "thisago"
           tdiv(class = "glory"):
             text "All glory to YAHUAH"
-
+        script(src = assetUrl "script/incAccess.js")
 
   resp(&"<!DOCTYPE html>\l{vnode}", view.code)